requirements.txt

fastjsonschema==2.17.1
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
fonttools==4.40.0
greenlet==2.0.2
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
PyMySQL==1.1.0
SQLAlchemy==2.0.17
nltk==3.8.1
numpy==1.23.5
pickleshare==0.7.5

Mako
MarkupSafe
mysqlclient==1.3.13
python-dateutil==2.7.4
python-dotenv==1.0.0
python-dotenv[cli]
python-editor==1.0.3
six==1.11.0
Werkzeug==2.2.2
